refactor: capitalize type name
