+++ packages/steps/src/style/description.tsx
@@ -1,15 +1,9 @@
 import { css } from "@emotion/react"
 import { globalColor, illaPrefix } from "@illa-design/theme"
 import { SerializedStyles } from "@emotion/serialize"
-import {
-  stepSize,
-  stepType,
-  labelPlacement,
-  StepStyleConfig,
-} from "../interface"
-import React, { CSSProperties } from "react"
+import { StepSize } from "../interface"
 
-export function applyDescriptionStyle(size: stepSize): SerializedStyles {
+export function applyDescriptionStyle(size: StepSize): SerializedStyles {
   return css([
     {
       color: globalColor(`--${illaPrefix}-gray-06`),
@@ -18,7 +12,7 @@ export function applyDescriptionStyle(size: stepSize): SerializedStyles {
   ])
 }
 
-function applyDescriptionSize(size: stepSize): SerializedStyles {
+function applyDescriptionSize(size: StepSize): SerializedStyles {
   let sizeCss: SerializedStyles
   switch (size) {
     default:
